Create WorkoutCard component and refactor HomePage
