+++ frontend/src/pages/HomePage.tsx
@@ -1,19 +1,15 @@
 import "./HomePage.css";
 import {useNavigate} from "react-router-dom";
 import {HomeProps, Workout} from "../types/types.ts";
-import RightArrow from "../components/svg/RightArrow.tsx";
 import {useContext} from "react";
 import {WorkoutContext} from "../components/state/WorkoutContext.tsx";
 import ProfileIcon from "../components/svg/ProfileIcon.tsx";
 import PlusIcon from "../components/svg/PlusIcon.tsx";
+import WorkoutCard from "../components/workoutcard/WorkoutCard.tsx";
 
 function HomePage({userName, imageUrl}:Readonly<HomeProps>) {
     const navigate = useNavigate();
     const {workouts} = useContext(WorkoutContext);
-
-    function handleClick(id:string) {
-        navigate(`/workout/${id}`);
-    }
 
     return(
         <>
@@ -41,28 +37,7 @@ function HomePage({userName, imageUrl}:Readonly<HomeProps>) {
                     <h2 className={"home-title"}>Recent Workouts</h2>
                     {workouts.map((workout:Workout) => {
                         return(
-                            <div key={workout.id} className={"workout-card"}>
-                                <div className={"workout-card-head"}>
-                                    <h3>{workout.name}</h3>
-                                    <p>{workout.day}</p>
-                                </div>
-                                <div className={"workout-card-body"}>
-                                    <p>{workout.description}</p>
-                                    <p>{workout.plan.length} Exercise(s)</p>
-                                </div>
-                                <button
-                                    className={"icon-bottom-right btn-details"}
-                                    onClick={() => workout.id && handleClick(workout.id)}
-                                    onKeyDown={(event) => {
-                                        if(event.key === "Enter") {
-                                            workout.id && handleClick(workout.id);
-                                        }
-                                    }}
-                                    aria-labelledby={`workoutName-${workout.id}`}
-                                >
-                                    <RightArrow />
-                                </button>
-                            </div>
+                            <WorkoutCard workout={workout} />
                         )
                     })}
                 </div>
